Implement multiple scenes
The game will have at least 3 scenes: title, dungeon, and gameover.
Right now only DungeonScene is implemented, but it is a good proof of
concept. Many classes have been changed to reference DungeonScene
instead of DarkMaze (all Entities). Drawing code specific to a Scene is
now in that Scene rather than in Screen.