Fix certain pathing cases for cloaks
If a cloak had line of sight to a firecracker when it went off, but was
already moving perpendicularly away from that tile, it would get the (!)
over its head but proceed to go in the wrong direction. This has been
corrected by allowing them to double back on themselves in this limited
case.

Also, cloak constructor takes x and y offsets instead of a bbox.